+++ src/gdt.rs
@@ -11,6 +11,12 @@ static GDT_ENTRIES: usize = 5;
 /// The loaded GDT.
 pub static GDT: InitLater<Gdt> = InitLater::uninit();
 
+/// The size of the emergency stack, in bytes.
+static STACK_SIZE: u64 = 4096 * 5;
+
+/// The emergency stack given to IST 1.
+static mut STACK: [u8; STACK_SIZE as usize] = [0; STACK_SIZE as usize];
+
 /// Offset in the GDT where the kernel's code segment will be.
 #[unsafe(no_mangle)]
 static CODE_SEGMENT_OFFSET: u16 = 0x8;
@@ -20,12 +26,13 @@ static TSS_SEGMENT_OFFSET: u64 = 0x18;
 
 /// The Global Descriptor Table.
 /// [`Reference`](https://wiki.osdev.org/Global_Descriptor_Table)
+#[derive(Debug)]
 #[repr(transparent)]
 pub struct Gdt([SegmentDescriptor; GDT_ENTRIES]);
 
 /// A segment descriptor in the GDT.
 /// [`Reference`](https://wiki.osdev.org/Global_Descriptor_Table#Segment_Descriptor)
-#[derive(Default)]
+#[derive(Debug, Default)]
 #[repr(transparent)]
 struct SegmentDescriptor(u64);
 
@@ -42,7 +49,7 @@ static TSS: InitLater<Tss> = InitLater::uninit();
 
 /// The 64 bit Task State Segment.
 /// [`Reference`](https://wiki.osdev.org/Task_State_Segment)
-#[derive(Default)]
+#[derive(Debug, Default)]
 #[repr(C, packed(4))]
 pub struct Tss {
     _reserved_1: u32,
@@ -89,7 +96,7 @@ struct SystemSegmentDescriptor {
 
 impl SystemSegmentDescriptor {
     /// Creates a new descriptor from the provided TSS.
-    fn new(tss: &'static Tss) -> Self {
+    fn new_tss(tss: &'static Tss) -> Self {
         /// Present, available 64 bit TSS
         static ACCESS: u8 = 0b1000_1001;
 
@@ -108,25 +115,19 @@ impl SystemSegmentDescriptor {
     }
 }
 
-/// Creates a new TSS into the `TSS` static.
-/// Gives the first IST stack pointer having it's own stack.
+/// Loads a new TSS into the `TSS` static.
+/// Gives the first IST stack pointer it's own stack.
 pub fn setup_tss() -> Result<(), InitError<Tss>> {
-    /// The size of the stack given to IST 1, in bytes.
-    static SIZE: u64 = 4096 * 5;
-
-    /// The stack given to IST 1.
-    static mut STACK: [u8; SIZE as usize] = [0; SIZE as usize];
-
+    // Calculate stack start & end addresses
     let mut tss = Tss::default();
     let stack_addr = &raw const STACK as u64;
-    let stack_end_addr = stack_addr + SIZE;
+    let stack_end_addr = stack_addr + STACK_SIZE;
     dbg_info!("emergency stack at 0x{stack_addr:x} to 0x{stack_end_addr:x}");
 
-    // Load TSS with new stack into static
+    // Load the TSS into it's static
     tss.ist[0] = stack_end_addr;
     tss.iomap = size_of::<Tss>() as u16;
     TSS.init(tss)?;
-
     dbg_info!("TSS at 0x{:x}", &raw const TSS as u64);
 
     Ok(())
@@ -134,8 +135,8 @@ pub fn setup_tss() -> Result<(), InitError<Tss>> {
 
 /// Loads the TSS into the task register.
 pub fn load_tss() -> Result<(), LoadRegisterError<Tss>> {
+    // Bail if no TSS or no GDT
     TSS.read()?;
-
     if !startup::gdt_init() {
         do yeet LoadRegisterError::Other("GDT is not initialised!!!")
     }
@@ -160,14 +161,14 @@ pub fn load_gdt() -> Result<(), LoadRegisterError<Gdt>> {
     interrupts::cli();
     let mut gdt = Gdt([const { SegmentDescriptor(0) }; GDT_ENTRIES]);
 
-    // Init GDT with a code & data segment
+    // Add a code & data segment
     gdt.0[1] = SegmentDescriptor::new(true); // Loaded at CODE_SEGMENT_OFFSET
     gdt.0[2] = SegmentDescriptor::new(false); // <- is this needed?
 
     // Add TSS descriptor
-    // Don't need to log an error if the read fails, as it would be printed in the 'Prepared TSS load' startup task
+    // Don't need to log an error if the read fails, since it would be printed in the 'Prepared TSS load' startup task
     if let Ok(tss) = TSS.read() {
-        let desc = SystemSegmentDescriptor::new(tss);
+        let desc = SystemSegmentDescriptor::new_tss(tss);
 
         // Safety: The gdt doesn't actually need these values to be segment descriptors,
         // two back to back can instead be a single system segment descriptor, like what we're doing here
@@ -175,7 +176,8 @@ pub fn load_gdt() -> Result<(), LoadRegisterError<Gdt>> {
             mem::transmute::<SystemSegmentDescriptor, (SegmentDescriptor, SegmentDescriptor)>(desc)
         };
 
-        gdt.0[3] = low; // Loaded at TSS_SEGMENT_OFFSET
+        // Load the TSS descriptor at TSS_SEGMENT_OFFSET
+        gdt.0[3] = low;
         gdt.0[4] = high;
     }
 
@@ -183,13 +185,10 @@ pub fn load_gdt() -> Result<(), LoadRegisterError<Gdt>> {
     let _gdt = GDT.init(gdt)?;
     dbg_info!("GDT loaded at 0x{:x}", _gdt as *const Gdt as u64);
 
-    // Descriptor to be loaded into GDTR
+    // Load the GDT to it's register
     let descriptor = TableDescriptor::new(GDT.read()?);
-
     // Safety: The GDT and it's descriptor MUST be valid by this point
-    unsafe {
-        asm!("lgdt ({0})", in(reg) &descriptor, options(att_syntax, nostack));
-    }
+    unsafe { asm!("lgdt ({0})", in(reg) &descriptor, options(att_syntax, nostack)) }
 
     if gdt_register() != descriptor {
         do yeet LoadRegisterError::Store("GDT");
@@ -212,6 +211,14 @@ pub fn gdt_register() -> TableDescriptor<Gdt> {
     gdt
 }
 
+/// Returns the current value in the Code Segment register.
+pub fn cs_register() -> u16 {
+    let cs;
+    // Safety: We're just copying over a register
+    unsafe { asm!("mov {0:x}, cs", out(reg) cs, options(preserves_flags, nostack)) }
+    cs
+}
+
 /// Reloads the CS register
 /// # Safety
 /// There must be a valid code segment where the `CODE_SEGMENT_OFFSET` static is pointing in the GDT.
@@ -226,5 +233,50 @@ unsafe extern "C" fn reload_cs() {
             addr = lateout(reg) _,
             options(preserves_flags),
         )
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    /// Tests that various structs passed to the CPU are the size that the CPU expects them.
+    #[test_case]
+    fn structs_have_the_right_size() {
+        let segment_size = size_of::<SegmentDescriptor>();
+        assert_eq!(size_of::<Tss>(), 104);
+        assert_eq!(segment_size, 8);
+        assert_eq!(size_of::<SystemSegmentDescriptor>(), segment_size * 2);
+        assert_eq!(size_of::<Gdt>(), segment_size * GDT_ENTRIES)
+    }
+
+    /// Tests that the CS register equals the `CODE_SEGMENT_OFFSET` static.
+    #[test_case]
+    fn cs_equals_static() {
+        GDT.read().unwrap(); // if the GDT isn't init, CS may not equal CODE_SEGMENT_OFFSET
+        assert_eq!(cs_register(), CODE_SEGMENT_OFFSET)
+    }
+
+    /// Tests that a TSS System Segment Descriptor actually points to the TSS.
+    #[test_case]
+    fn tss_segment_has_correct_ptr() {
+        let tss = TSS.read().unwrap();
+        let ptr = tss as *const Tss as u64;
+        let segment = SystemSegmentDescriptor::new_tss(tss);
+
+        let mut segment_ptr = segment.offset_very_low as u64;
+        segment_ptr |= (segment.offset_low as u64) << 16;
+        segment_ptr |= (segment.offset_medium as u64) << 24;
+        segment_ptr |= (segment.offset_high as u64) << 32;
+        assert_eq!(ptr, segment_ptr)
+    }
+
+    /// Tests that IST 1 points to the emergency stack.
+    #[test_case]
+    fn ist_one_points_to_stack() {
+        let tss = TSS.read().unwrap();
+        let stack_end_addr = &raw const STACK as u64 + STACK_SIZE;
+        let ist1 = tss.ist[0];
+        assert_eq!(ist1, stack_end_addr)
     }
 }